requirements.txt
pillow==10.4.0
plotly==6.2.0
pyPDF2==3.0.1
pydantic==2.11.7
pytesseract==0.3.13
opencv-python==4.12.0.88
SQLAlchemy==2.0.41
numpy==2.2.6
pandas==2.3.1
streamlit==1.47.0
matplotlib==3.10.3
seaborn==0.13.2
passlib[bcrypt]==1.7.4
